+++ public/assets/css/blogcard.scss
@@ -1,13 +1,13 @@
-.slide-in {
-  transform: translateX(0%);
-  opacity: 1;
-  transition: transform 0.5s, opacity 0.5s;
+.ccard{
+  margin-left: 27vh;
+  height: 35vh;
+  @media screen and (max-width:600px) {
+      margin-left: -3vh;
+      height: 30vh;
+      width: 40vh;
+  }
 }
 
-.slide-out {
-  transform: translateX(100%);
-  opacity: 0;
-}
 .card-container {
     max-height: 500px; /* Adjust the max height as needed */
     // overflow-y: auto;
