--- public/assets/css/projects.scss
@@ -1,54 +1,0 @@
-/* styles/services.scss */
-
-.services-page {
-  text-align: center;
-  padding: 20px;
-  margin-top: 100px;
-  h1 {
-    font-size: 36px;
-    margin-bottom: 20px;
-  }
-
-  .service-cards {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-
-    .service-card {
-      border: 1px solid #ddd;
-      border-radius: 5px;
-      margin: 10px;
-      width: 300px;
-      // background-color: #fff;
-      box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
-      overflow: hidden;
-
-      .card-image {
-        width: 100%;
-        height: 150px;
-        background-size: cover;
-        background-position: center;
-        transition: transform 0.3s ease; // Add a transform transition for hover effect
-      }
-
-      .card-content {
-        padding: 20px;
-
-        h2 {
-          font-size: 24px;
-          margin-bottom: 10px;
-        }
-
-        p {
-          font-size: 16px;
-        }
-      }
-
-      &:hover {
-        .card-image {
-          transform: scale(1.1); // Increase size on hover
-        }
-      }
-    }
-  }
-}
