--- public/assets/css/snowfall.scss
@@ -1,34 +1,0 @@
-/* Snowfall.css */
-
-/* Define the snowflake */
-.snowfall-container {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    pointer-events: none;
-    z-index: 1000;
-    overflow: hidden;
-  }
-.snowflake {
-    position: absolute;
-    width: 5px;
-    height: 5px;
-    background-color: #fff;
-    border-radius: 50%;
-    pointer-events: none;
-    animation: snowfall 5s linear infinite;
-  }
-  
-  /* Keyframes for the snowfall animation */
-  @keyframes snowfall {
-    0% {
-      transform: translateY(-10vh) translateX(-10px) scale(0.8);
-    }
-    100% {
-      transform: translateY(110vh) translateX(20px) scale(1.2);
-      opacity: 0;
-    }
-  }
-
